Update root directory
Root directory was previously being set based on the instructions
of setting up the api server, which meant that it was "src".
By running the api server in root directory "iac-scan-runner"
instead of in "src", the relative paths become simpler
and tests are easier to run on local machines.

Root directory is now set to "iac-scan-runnner" instead of
"iac-scan-runner/src".